+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporterOptions.cs
@@ -109,9 +109,14 @@ namespace OpenTelemetry.Exporter
         public BatchExportProcessorOptions<Activity> BatchExportProcessorOptions { get; set; } = new BatchExportActivityProcessorOptions();
 
         /// <summary>
-        /// Gets or sets the metric export interval in milliseconds. The default value is 60000.
+        /// Gets or sets the <see cref="MetricReaderType" /> to use. Defaults to <c>MetricReaderType.Periodic</c>.
         /// </summary>
-        public int MetricExportIntervalMilliseconds { get; set; } = 60000;
+        public MetricReaderType MetricReaderType { get; set; } = MetricReaderType.Periodic;
+
+        /// <summary>
+        /// Gets or sets the <see cref="PeriodicExportingMetricReaderOptions" /> options. Ignored unless <c>MetricReaderType</c> is <c>Periodic</c>.
+        /// </summary>
+        public PeriodicExportingMetricReaderOptions PeriodicExportingMetricReaderOptions { get; set; } = new PeriodicExportingMetricReaderOptions();
 
         /// <summary>
         /// Gets or sets the AggregationTemporality used for Histogram