+++ src/OpenTelemetry/Metrics/BaseExportingMetricReader.cs
@@ -38,8 +38,7 @@ namespace OpenTelemetry.Metrics
             if (attributes.Length > 0)
             {
                 var attr = (AggregationTemporalityAttribute)attributes[attributes.Length - 1];
-                this.PreferredAggregationTemporality = attr.Preferred;
-                this.SupportedAggregationTemporality = attr.Supported;
+                this.Temporality = attr.Temporality;
             }
 
             attributes = exportorType.GetCustomAttributes(typeof(ExportModesAttribute), true);
@@ -79,7 +78,7 @@ namespace OpenTelemetry.Metrics
         }
 
         /// <inheritdoc/>
-        protected override bool ProcessMetrics(in Batch<Metric> metrics, int timeoutMilliseconds)
+        internal override bool ProcessMetrics(in Batch<Metric> metrics, int timeoutMilliseconds)
         {
             // TODO: Do we need to consider timeout here?
             return this.exporter.Export(metrics) == ExportResult.Success;