+++ src/OpenTelemetry/Metrics/MetricPoint.cs
@@ -22,12 +22,17 @@ namespace OpenTelemetry.Metrics
 {
     public struct MetricPoint
     {
+        internal DateTimeOffset StartTime;
+        internal DateTimeOffset EndTime;
+
         private readonly AggregationType aggType;
-        private readonly HistogramMeasurements histogramMeasurements;
-        private long longVal;
-        private long lastLongSum;
-        private double doubleVal;
-        private double lastDoubleSum;
+        private readonly HistogramBuckets histogramBuckets;
+
+        // Represents either "value" for double/long metric types or "count" when histogram
+        private MetricPointValueStorage primaryValue;
+
+        // Represents either "lastValue" for double/long metric types when delta or "sum" when histogram
+        private MetricPointValueStorage secondaryValue;
 
         internal MetricPoint(
             AggregationType aggType,
@@ -42,25 +47,21 @@ namespace OpenTelemetry.Metrics
             this.StartTime = startTime;
             this.Tags = new ReadOnlyTagCollection(keys, values);
             this.EndTime = default;
-            this.LongValue = default;
-            this.longVal = default;
-            this.lastLongSum = default;
-            this.DoubleValue = default;
-            this.doubleVal = default;
-            this.lastDoubleSum = default;
+            this.primaryValue = default;
+            this.secondaryValue = default;
             this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
             if (this.aggType == AggregationType.Histogram)
             {
-                this.histogramMeasurements = new HistogramMeasurements(histogramBounds);
+                this.histogramBuckets = new HistogramBuckets(histogramBounds);
             }
             else if (this.aggType == AggregationType.HistogramSumCount)
             {
-                this.histogramMeasurements = new HistogramMeasurements(null);
+                this.histogramBuckets = new HistogramBuckets(null);
             }
             else
             {
-                this.histogramMeasurements = null;
+                this.histogramBuckets = null;
             }
         }
 
@@ -69,37 +70,57 @@ namespace OpenTelemetry.Metrics
         /// </summary>
         public ReadOnlyTagCollection Tags { get; }
 
-        public DateTimeOffset StartTime { get; internal set; }
-
-        public DateTimeOffset EndTime { get; internal set; }
-
-        public long LongValue { get; internal set; }
-
-        public double DoubleValue { get; internal set; }
-
         internal MetricPointStatus MetricPointStatus { get; private set; }
 
-        public long[] GetBucketCounts()
+        public DateTimeOffset GetStartTime() => this.StartTime;
+
+        public DateTimeOffset GetEndTime() => this.EndTime;
+
+        public long GetSumLong()
         {
-            if (this.aggType == AggregationType.Histogram)
+            if (this.aggType == AggregationType.LongSumIncomingDelta || this.aggType == AggregationType.LongSumIncomingCumulative)
             {
-                return this.histogramMeasurements.AggregatedBucketCounts;
+                return this.primaryValue.SnapshotAsLong;
             }
             else
             {
-                throw new NotSupportedException($"{nameof(this.GetBucketCounts)} is not supported for this metric type.");
+                throw new NotSupportedException($"{nameof(this.GetSumLong)} is not supported for this metric type.");
             }
         }
 
-        public double[] GetExplicitBounds()
+        public double GetSumDouble()
         {
-            if (this.aggType == AggregationType.Histogram)
+            if (this.aggType == AggregationType.DoubleSumIncomingDelta || this.aggType == AggregationType.DoubleSumIncomingCumulative)
             {
-                return this.histogramMeasurements.ExplicitBounds;
+                return this.primaryValue.SnapshotAsDouble;
             }
             else
             {
-                throw new NotSupportedException($"{nameof(this.GetExplicitBounds)} is not supported for this metric type.");
+                throw new NotSupportedException($"{nameof(this.GetSumDouble)} is not supported for this metric type.");
+            }
+        }
+
+        public long GetGaugeLastValueLong()
+        {
+            if (this.aggType == AggregationType.LongGauge)
+            {
+                return this.primaryValue.SnapshotAsLong;
+            }
+            else
+            {
+                throw new NotSupportedException($"{nameof(this.GetGaugeLastValueLong)} is not supported for this metric type.");
+            }
+        }
+
+        public double GetGaugeLastValueDouble()
+        {
+            if (this.aggType == AggregationType.DoubleGauge)
+            {
+                return this.primaryValue.SnapshotAsDouble;
+            }
+            else
+            {
+                throw new NotSupportedException($"{nameof(this.GetGaugeLastValueDouble)} is not supported for this metric type.");
             }
         }
 
@@ -107,7 +128,7 @@ namespace OpenTelemetry.Metrics
         {
             if (this.aggType == AggregationType.Histogram || this.aggType == AggregationType.HistogramSumCount)
             {
-                return this.histogramMeasurements.Count;
+                return this.primaryValue.SnapshotAsLong;
             }
             else
             {
@@ -119,11 +140,23 @@ namespace OpenTelemetry.Metrics
         {
             if (this.aggType == AggregationType.Histogram || this.aggType == AggregationType.HistogramSumCount)
             {
-                return this.histogramMeasurements.Sum;
+                return this.secondaryValue.SnapshotAsDouble;
             }
             else
             {
                 throw new NotSupportedException($"{nameof(this.GetHistogramSum)} is not supported for this metric type.");
+            }
+        }
+
+        public HistogramBuckets GetHistogramBuckets()
+        {
+            if (this.aggType == AggregationType.Histogram || this.aggType == AggregationType.HistogramSumCount)
+            {
+                return this.histogramBuckets;
+            }
+            else
+            {
+                throw new NotSupportedException($"{nameof(this.GetHistogramBuckets)} is not supported for this metric type.");
             }
         }
 
@@ -133,19 +166,19 @@ namespace OpenTelemetry.Metrics
             {
                 case AggregationType.LongSumIncomingDelta:
                     {
-                        Interlocked.Add(ref this.longVal, number);
+                        Interlocked.Add(ref this.primaryValue.CurrentAsLong, number);
                         break;
                     }
 
                 case AggregationType.LongSumIncomingCumulative:
                     {
-                        Interlocked.Exchange(ref this.longVal, number);
+                        Interlocked.Exchange(ref this.primaryValue.CurrentAsLong, number);
                         break;
                     }
 
                 case AggregationType.LongGauge:
                     {
-                        Interlocked.Exchange(ref this.longVal, number);
+                        Interlocked.Exchange(ref this.primaryValue.CurrentAsLong, number);
                         break;
                     }
 
@@ -180,42 +213,42 @@ namespace OpenTelemetry.Metrics
                         double initValue, newValue;
                         do
                         {
-                            initValue = this.doubleVal;
+                            initValue = this.primaryValue.CurrentAsDouble;
                             newValue = initValue + number;
                         }
-                        while (initValue != Interlocked.CompareExchange(ref this.doubleVal, newValue, initValue));
+                        while (initValue != Interlocked.CompareExchange(ref this.primaryValue.CurrentAsDouble, newValue, initValue));
                         break;
                     }
 
                 case AggregationType.DoubleSumIncomingCumulative:
                     {
-                        Interlocked.Exchange(ref this.doubleVal, number);
+                        Interlocked.Exchange(ref this.primaryValue.CurrentAsDouble, number);
                         break;
                     }
 
                 case AggregationType.DoubleGauge:
                     {
-                        Interlocked.Exchange(ref this.doubleVal, number);
+                        Interlocked.Exchange(ref this.primaryValue.CurrentAsDouble, number);
                         break;
                     }
 
                 case AggregationType.Histogram:
                     {
                         int i;
-                        for (i = 0; i < this.histogramMeasurements.ExplicitBounds.Length; i++)
+                        for (i = 0; i < this.histogramBuckets.ExplicitBounds.Length; i++)
                         {
                             // Upper bound is inclusive
-                            if (number <= this.histogramMeasurements.ExplicitBounds[i])
+                            if (number <= this.histogramBuckets.ExplicitBounds[i])
                             {
                                 break;
                             }
                         }
 
-                        lock (this.histogramMeasurements.LockObject)
+                        lock (this.histogramBuckets.LockObject)
                         {
-                            this.histogramMeasurements.CountVal++;
-                            this.histogramMeasurements.SumVal += number;
-                            this.histogramMeasurements.BucketCounts[i]++;
+                            this.primaryValue.CurrentAsLong++;
+                            this.secondaryValue.CurrentAsDouble += number;
+                            this.histogramBuckets.BucketCounts[i]++;
                         }
 
                         break;
@@ -223,10 +256,10 @@ namespace OpenTelemetry.Metrics
 
                 case AggregationType.HistogramSumCount:
                     {
-                        lock (this.histogramMeasurements.LockObject)
+                        lock (this.histogramBuckets.LockObject)
                         {
-                            this.histogramMeasurements.CountVal++;
-                            this.histogramMeasurements.SumVal += number;
+                            this.primaryValue.CurrentAsLong++;
+                            this.secondaryValue.CurrentAsDouble += number;
                         }
 
                         break;
@@ -256,21 +289,21 @@ namespace OpenTelemetry.Metrics
                     {
                         if (outputDelta)
                         {
-                            long initValue = Interlocked.Read(ref this.longVal);
-                            this.LongValue = initValue - this.lastLongSum;
-                            this.lastLongSum = initValue;
+                            long initValue = Interlocked.Read(ref this.primaryValue.CurrentAsLong);
+                            this.primaryValue.SnapshotAsLong = initValue - this.secondaryValue.CurrentAsLong;
+                            this.secondaryValue.CurrentAsLong = initValue;
                             this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                             // Check again if value got updated, if yes reset status.
                             // This ensures no Updates get Lost.
-                            if (initValue != Interlocked.Read(ref this.longVal))
+                            if (initValue != Interlocked.Read(ref this.primaryValue.CurrentAsLong))
                             {
                                 this.MetricPointStatus = MetricPointStatus.CollectPending;
                             }
                         }
                         else
                         {
-                            this.LongValue = Interlocked.Read(ref this.longVal);
+                            this.primaryValue.SnapshotAsLong = Interlocked.Read(ref this.primaryValue.CurrentAsLong);
                         }
 
                         break;
@@ -286,14 +319,14 @@ namespace OpenTelemetry.Metrics
                             // As long as the value is not -ve infinity,
                             // the exchange (to 0.0) will never occur,
                             // but we get the original value atomically.
-                            double initValue = Interlocked.CompareExchange(ref this.doubleVal, 0.0, double.NegativeInfinity);
-                            this.DoubleValue = initValue - this.lastDoubleSum;
-                            this.lastDoubleSum = initValue;
+                            double initValue = Interlocked.CompareExchange(ref this.primaryValue.CurrentAsDouble, 0.0, double.NegativeInfinity);
+                            this.primaryValue.SnapshotAsDouble = initValue - this.secondaryValue.CurrentAsDouble;
+                            this.secondaryValue.CurrentAsDouble = initValue;
                             this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                             // Check again if value got updated, if yes reset status.
                             // This ensures no Updates get Lost.
-                            if (initValue != Interlocked.CompareExchange(ref this.doubleVal, 0.0, double.NegativeInfinity))
+                            if (initValue != Interlocked.CompareExchange(ref this.primaryValue.CurrentAsDouble, 0.0, double.NegativeInfinity))
                             {
                                 this.MetricPointStatus = MetricPointStatus.CollectPending;
                             }
@@ -305,7 +338,7 @@ namespace OpenTelemetry.Metrics
                             // As long as the value is not -ve infinity,
                             // the exchange (to 0.0) will never occur,
                             // but we get the original value atomically.
-                            this.DoubleValue = Interlocked.CompareExchange(ref this.doubleVal, 0.0, double.NegativeInfinity);
+                            this.primaryValue.SnapshotAsDouble = Interlocked.CompareExchange(ref this.primaryValue.CurrentAsDouble, 0.0, double.NegativeInfinity);
                         }
 
                         break;
@@ -313,12 +346,12 @@ namespace OpenTelemetry.Metrics
 
                 case AggregationType.LongGauge:
                     {
-                        this.LongValue = Interlocked.Read(ref this.longVal);
+                        this.primaryValue.SnapshotAsLong = Interlocked.Read(ref this.primaryValue.CurrentAsLong);
                         this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                         // Check again if value got updated, if yes reset status.
                         // This ensures no Updates get Lost.
-                        if (this.LongValue != Interlocked.Read(ref this.longVal))
+                        if (this.primaryValue.SnapshotAsLong != Interlocked.Read(ref this.primaryValue.CurrentAsLong))
                         {
                             this.MetricPointStatus = MetricPointStatus.CollectPending;
                         }
@@ -333,12 +366,12 @@ namespace OpenTelemetry.Metrics
                         // As long as the value is not -ve infinity,
                         // the exchange (to 0.0) will never occur,
                         // but we get the original value atomically.
-                        this.DoubleValue = Interlocked.CompareExchange(ref this.doubleVal, 0.0, double.NegativeInfinity);
+                        this.primaryValue.SnapshotAsDouble = Interlocked.CompareExchange(ref this.primaryValue.CurrentAsDouble, 0.0, double.NegativeInfinity);
                         this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                         // Check again if value got updated, if yes reset status.
                         // This ensures no Updates get Lost.
-                        if (this.DoubleValue != Interlocked.CompareExchange(ref this.doubleVal, 0.0, double.NegativeInfinity))
+                        if (this.primaryValue.SnapshotAsDouble != Interlocked.CompareExchange(ref this.primaryValue.CurrentAsDouble, 0.0, double.NegativeInfinity))
                         {
                             this.MetricPointStatus = MetricPointStatus.CollectPending;
                         }
@@ -348,22 +381,22 @@ namespace OpenTelemetry.Metrics
 
                 case AggregationType.Histogram:
                     {
-                        lock (this.histogramMeasurements.LockObject)
+                        lock (this.histogramBuckets.LockObject)
                         {
-                            this.histogramMeasurements.Count = this.histogramMeasurements.CountVal;
-                            this.histogramMeasurements.Sum = this.histogramMeasurements.SumVal;
+                            this.primaryValue.SnapshotAsLong = this.primaryValue.CurrentAsLong;
+                            this.secondaryValue.SnapshotAsDouble = this.secondaryValue.CurrentAsDouble;
                             if (outputDelta)
                             {
-                                this.histogramMeasurements.CountVal = 0;
-                                this.histogramMeasurements.SumVal = 0;
+                                this.primaryValue.CurrentAsLong = 0;
+                                this.secondaryValue.CurrentAsDouble = 0;
                             }
 
-                            for (int i = 0; i < this.histogramMeasurements.BucketCounts.Length; i++)
+                            for (int i = 0; i < this.histogramBuckets.BucketCounts.Length; i++)
                             {
-                                this.histogramMeasurements.AggregatedBucketCounts[i] = this.histogramMeasurements.BucketCounts[i];
+                                this.histogramBuckets.AggregatedBucketCounts[i] = this.histogramBuckets.BucketCounts[i];
                                 if (outputDelta)
                                 {
-                                    this.histogramMeasurements.BucketCounts[i] = 0;
+                                    this.histogramBuckets.BucketCounts[i] = 0;
                                 }
                             }
 
@@ -375,14 +408,14 @@ namespace OpenTelemetry.Metrics
 
                 case AggregationType.HistogramSumCount:
                     {
-                        lock (this.histogramMeasurements.LockObject)
+                        lock (this.histogramBuckets.LockObject)
                         {
-                            this.histogramMeasurements.Count = this.histogramMeasurements.CountVal;
-                            this.histogramMeasurements.Sum = this.histogramMeasurements.SumVal;
+                            this.primaryValue.SnapshotAsLong = this.primaryValue.CurrentAsLong;
+                            this.secondaryValue.SnapshotAsDouble = this.secondaryValue.CurrentAsDouble;
                             if (outputDelta)
                             {
-                                this.histogramMeasurements.CountVal = 0;
-                                this.histogramMeasurements.SumVal = 0;
+                                this.primaryValue.CurrentAsLong = 0;
+                                this.secondaryValue.CurrentAsDouble = 0;
                             }
 
                             this.MetricPointStatus = MetricPointStatus.NoCollectPending;