+++ src/OpenTelemetry/Metrics/MetricTypeExtensions.cs
@@ -26,8 +26,8 @@ namespace OpenTelemetry.Metrics
 
         internal const MetricType METRIC_TYPE_SUM = (MetricType)0x10;
         internal const MetricType METRIC_TYPE_GAUGE = (MetricType)0x20;
-        internal const MetricType METRIC_TYPE_HISTOGRAM = (MetricType)0x30;
-        /* internal const byte METRIC_TYPE_SUMMARY = 0x40; // not used */
+        /* internal const byte METRIC_TYPE_SUMMARY = 0x30; // not used */
+        internal const MetricType METRIC_TYPE_HISTOGRAM = (MetricType)0x40;
 
         internal const MetricType POINT_KIND_MASK = (MetricType)0x0f;
 
@@ -59,7 +59,7 @@ namespace OpenTelemetry.Metrics
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsHistogram(this MetricType self)
         {
-            return (self & METRIC_TYPE_MASK) == METRIC_TYPE_HISTOGRAM;
+            return self.HasFlag(METRIC_TYPE_HISTOGRAM);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]