+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Added configuration options for `MetricReaderType` to allow for configuring
+  the `OtlpMetricExporter` to export either manually or periodically.
+  ([#2674](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2674))
+* The internal log message used when OTLP export client connection failure occurs,
+  will now include the endpoint uri as well.
+  ([#2686](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2686))
+
 ## 1.2.0-beta2
 
 Released 2021-Nov-19